split ci package
