+++ ci/ci.go
@@ -1,4 +1,4 @@
-package main
+package ci
 
 import (
 	"fmt"
