+++ ci/ci_test.go
@@ -1,4 +1,4 @@
-package main
+package ci
 
 import (
 	"os"
